test/web-platform-tests/to-upstream/xhr/access-control-preflight-request-header-returns-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Access-Control-Request-Origin accept different origin between preflight and actual request</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
      "use strict";

      // duplicated from wpt/tests/cors/support.js
      function corsUrl(path) {
        function dirname(dirPath) {
          return dirPath.replace(/\/[^/]*$/, "/");
        }

        /* This subdomain should point to this same location */
        const SUBDOMAIN = "www1";

        /* Changes http://example.com/abc/def/cool.htm to http://www1.example.com/abc/def/ */
        const CROSSDOMAIN = dirname(location.href).replace("://", "://" + SUBDOMAIN + ".");

        const url = CROSSDOMAIN + path;
        return url
      }

      async_test(function(test) {
        const xhr = new XMLHttpRequest;

        xhr.open("GET", corsUrl("resources/access-control-preflight-request-header-returns-origin.py"));

        xhr.setRequestHeader("X-Test", "foobar");

        xhr.onerror = test.unreached_func("Error occurred.");

        xhr.onload = test.step_func_done(function() {
          assert_equals(xhr.status, 200);
          assert_equals(xhr.responseText, "PASS");
        });

        xhr.send();
      }, "Receives different origin values from preflight and actual request");
    </script>
  </body>
</html>
